# policy: improve performance of EndpointSelector Matches