When performing network policy scale testing in #40227 I've noticed that ~20% of object are allocated in (*EndpointSelector).Matches

This PR reduces allocations by half.

Before:
```
BenchmarkMatchesValid1000
BenchmarkMatchesValid1000-32              	100000000	       179.0 ns/op	      88 B/op	       2 allocs/op
BenchmarkMatchesInvalid1000
BenchmarkMatchesInvalid1000-32            	100000000	       214.2 ns/op	      88 B/op	       2 allocs/op
BenchmarkMatchesValid1000Parallel
BenchmarkMatchesValid1000Parallel-32      	100000000	        32.35 ns/op	      88 B/op	       2 allocs/op
BenchmarkMatchesInvalid1000Parallel
BenchmarkMatchesInvalid1000Parallel-32    	100000000	        33.35 ns/op	      88 B/op	       2 allocs/op
```

After:
```
BenchmarkMatchesValid1000
BenchmarkMatchesValid1000-32              	100000000	        85.30 ns/op	      24 B/op	       1 allocs/op
BenchmarkMatchesInvalid1000
BenchmarkMatchesInvalid1000-32            	100000000	        81.65 ns/op	      24 B/op	       1 allocs/op
BenchmarkMatchesValid1000Parallel
BenchmarkMatchesValid1000Parallel-32      	100000000	        27.45 ns/op	      24 B/op	       1 allocs/op
BenchmarkMatchesInvalid1000Parallel
BenchmarkMatchesInvalid1000Parallel-32    	100000000	        27.41 ns/op	      24 B/op	       1 allocs/op
```