@gandro 2025-07-15
Since this seems very low-risk, should we backport this to v1.18?