@jrajahalme 2025-07-16
> Since this seems very low-risk, should we backport this to v1.18?

Added the `needs-backport/1.18` label